Project Added
Ok, guys. So this is a working version. And I made a ton of changes to
the box stuff, so now all the enemies, walls, and the player, each have
one box that they all share. Hopefully this should solve the memory
problem. Let me know if you run into any issues.